+++ CHANGELOG.md
@@ -11,8 +11,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 20YY-MM-DD
+## [1.0.0] - 1/11/2026
 
 ### Added
+- Release
 
-- This CHANGELOG file to hopefully serve as an evolving example of a standardized open source project CHANGELOG.
+
+## [1.0.1] - 1/11/2026
+
+- lowkey dont remember
+
+## [1.0.2] - 1/11/2026
+## Added
+- Weight slider per source
+- Offset position/rotation toggle
+
+## Removed
+- "Add constraint w/offfset. Replaced with settings toggle